+++ plugins/postcss-design-tokens/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 - Added `valueFunctionName` option to control the `design-token` function name.
+- Added `atRuleName` option to control the `design-tokens` at-rule name.
 
 ### 1.0.0 (May 23, 2022)