+++ plugins/postcss-design-tokens/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changes to PostCSS Design Tokens
 
-### Unreleased
+### 1.1.0 (Aug 2, 2022)
 
 - Added `valueFunctionName` option to control the `design-token` function name.
 - Added `importAtRuleName` option to control the `design-tokens` at-rule name.